+++ code/sys.driver/driver-vite/src/common/libs.ts
@@ -5,8 +5,8 @@ export { Denofile } from '@sys/driver-deno/runtime';
 export { asArray, R, slug, Str, Time } from '@sys/std';
 
 export { c, Cli, stripAnsi } from '@sys/cli';
-export { Hash } from '@sys/crypto/hash';
 export { HashFmt } from '@sys/crypto/fmt';
+export { Hash } from '@sys/crypto/hash';
 export { Fs, Path, Pkg } from '@sys/fs';
-export { Process } from '@sys/proc';
+export { Process } from '@sys/process';
 export { Net } from '@sys/std-s/network';
