+++ code/sys.driver/driver-vitepress/src/common/t.ts
@@ -3,7 +3,7 @@ export type * from '@sys/types/t';
 
 export type { DenofileJson } from '@sys/driver-deno/t';
 export type { DirSnapshot, FileMap, FsPathFilter } from '@sys/fs/t';
-export type { ProcHandle, ProcReadySignalFilter } from '@sys/proc/t';
+export type { ProcHandle, ProcReadySignalFilter } from '@sys/process/t';
 export type { Ignore } from '@sys/std/t';
 export type { Tmpl, TmplFileOperation } from '@sys/tmpl/t';
