+++ code/sys/process/deno.json
@@ -1,5 +1,5 @@
 {
-  "name": "@sys/proc",
+  "name": "@sys/process",
   "version": "0.0.26",
   "license": "MIT",
   "tasks": {
