+++ code/sys/process/scripts/-clean.ts
