+++ code/sys/process/src/m.Process/u.kill.ts
