+++ code/sys/std-s/src/m.Server.Http/u.keyboard.ts
@@ -1,4 +1,4 @@
-import { Process } from '@sys/proc';
+import { Process } from '@sys/process';
 import { Cli, c } from './common.ts';
 
 /**
