+++ deno.json
@@ -28,7 +28,7 @@
     "./code/sys/testing",
     "./code/sys/fs",
     "./code/sys/cli",
-    "./code/sys/proc",
+    "./code/sys/process",
     "./code/sys/crypto",
 
     "./code/sys/std-s",
