+++ scripts/common.ts
@@ -3,7 +3,7 @@ export { Semver } from '@sys/std/semver';
 
 export { c, Cli } from '@sys/cli';
 export { Fs } from '@sys/fs';
-export { Process } from '@sys/proc';
+export { Process } from '@sys/process';
 
 export * as t from './t.ts';
 export { Path, Paths } from './u.paths.ts';
